+++ app-example/app/(tabs)/explore.tsx
