+++ app-example/app/(tabs)/index.tsx
