+++ app-example/components/Collapsible.tsx
