+++ app-example/components/HapticTab.tsx
