+++ app-example/components/ui/TabBarBackground.ios.tsx
