+++ app-example/components/ui/TabBarBackground.tsx
