+++ app-example/hooks/useColorScheme.ts
