+++ app-example/scripts/reset-project.js
